requirements.txt
torch
torchvision
basicsr
realesrgan
opencv-python
streamlit
numpy
Pillow